refactor: preQuestion 관련 타입 수정 반영
