+++ frontend/src/hooks/preQuestion/usePreQuestionModal.ts
@@ -2,14 +2,18 @@ import { useState } from 'react';
 
 import { useMutation } from '@tanstack/react-query';
 
-import { PreQuestionCustomHookType } from 'hooks/preQuestion/types/index';
-
 import useModal from 'hooks/useModal';
 import useSnackbar from 'hooks/utils/useSnackbar';
 
-import { requestGetPreQuestion } from 'apis/preQuestion';
+import { MESSAGE } from 'constants/constants';
+
+import {
+  PreQuestionDeleteRequestType,
+  requestDeletePreQuestion,
+  requestGetPreQuestion,
+} from 'apis/preQuestion';
+import { ParticipantType } from 'types/index';
 import { PreQuestionParticipantType } from 'types/preQuestion';
-import { ParticipantType } from 'types/team';
 
 const usePreQuestionModal = () => {
   const { isModalOpen, onClickOpenModal, onClickCloseModal } = useModal();
@@ -38,13 +42,13 @@ const usePreQuestionModal = () => {
   const onClickDeletePreQuestion = async ({
     levellogId,
     preQuestionId,
-  }: Pick<PreQuestionCustomHookType, 'levellogId' | 'preQuestionId'>) => {
+  }: Omit<PreQuestionDeleteRequestType, 'accessToken'>) => {
     // await deletePreQuestion({
     //   levellogId,
     //   preQuestionId,
     // });
-    // onClickCloseModal();
-    // showSnackbar({ message: MESSAGE.PREQUESTION_DELETE });
+    onClickCloseModal();
+    showSnackbar({ message: MESSAGE.PREQUESTION_DELETE });
   };
 
   return {
